+++ CHANGELOG.md
@@ -5,9 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.0] - ?
+## [1.2.0] - 2022-11-28
 
 - lock coordinates to valid map positions
+- prevent inbox with broken lock from processing
 - update inbox sync to coroutine for performance
 
 ## [1.1.0] - 2022-11-27